update create posts migration to enable cascade option when dropping tables
