+++ db/models/companies.js
@@ -1,0 +1,10 @@
+const db = require('../');
+
+const Company = db.Model.extend({
+  tableName: 'companies',
+  location: function() {
+    return this.hasOne('Location');
+  }
+});
+
+module.exports = db.model('Company', Company);
