+++ db/models/locations.js
@@ -1,0 +1,10 @@
+const db = require('../');
+
+const Location = db.Model.extend({
+  tableName: 'location',
+  company: function() {
+    return this.belongsTo('Company');
+  }
+});
+
+module.exports = db.model('Location', Location);
