+++ db/seeds/03_cards_seed.js
@@ -1,10 +1,8 @@
-//const models = require('../models');
-
 exports.seed = function(knex, Promise) {
-  // Deletes ALL existing entries
+
   return knex('cards').del()
     .then(function () {
-      // Inserts seed entries
+
       return knex('cards').insert([
         { id: 1,
           position: 'UI Developer',
@@ -13,13 +11,13 @@ exports.seed = function(knex, Promise) {
           notes: null,
           company_id: 1,
           user_id: 1
-        }
+        },
         { id: 2,
           position: 'Front End Developer',
           position_url: 'https://www.linkedin.com/jobs/view/430047268/',
           description: 'RHLA Technology is a leading provider of technology professionals on a project and full-time basis.',
           notes: 'Recruiter reached out to me.',
-          company_id: 3,
+          company_id: 2,
           user_id: 1
         },
         { id: 3,
@@ -27,7 +25,7 @@ exports.seed = function(knex, Promise) {
           position_url: 'https://www.linkedin.com/jobs/view/380992731/',
           description: 'Ladeeda is the leader in cloud security. ',
           notes: null,
-          company_id: 2,
+          company_id: 3,
           user_id: 1
         }
       ]);
